modify: move Vertex/Edge Query from utils to infrastructure package
